Use string function spects for citeproc_retrievers
Enables lazy import of citation_retreiver functions.
